# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "description": "API Documentation",
        "termsOfService": "",
        "contact": {
            "email": "ntam444@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get category",
                "description": "",
                "operationId": "category",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "the category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "code": "",
                                    "message": "",
                                    "error": "",
                                    "category": "{object}"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get list product",
                "description": "",
                "operationId": "products",
                "parameters": [
                    {
                        "parameter": "search_keyword",
                        "name": "search_keyword",
                        "in": "query",
                        "description": "the key word search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "sư"
                    },
                    {
                        "parameter": "category_id",
                        "name": "category_id",
                        "in": "query",
                        "description": "the category id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    },
                    {
                        "parameter": "store_id",
                        "name": "store_id",
                        "in": "query",
                        "description": "the store id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "code": "",
                                    "message": "",
                                    "error": "",
                                    "products": "array({object})"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get product",
                "description": "",
                "operationId": "product",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "the product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "code": "",
                                    "message": "",
                                    "error": "",
                                    "product": "{object}"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}